feat: combineAll <directory to pdf files>
The combineAll function allows the user to specify a directory. It will search the directory
and find all files that end with .pdf. It will show the current order of how the pdfs will
combine, and asks the user if they want to re-order it. If so, they can specify the order